chore: move sequencer config
