+++ crates/tempo-zone/src/rpc_client.rs
@@ -1,0 +1,9 @@
+use std::time::Duration;
+
+use alloy_rpc_client::ConnectionConfig;
+
+pub(crate) fn rpc_connection_config(retry_connection_interval: Duration) -> ConnectionConfig {
+    ConnectionConfig::new()
+        .with_max_retries(u32::MAX)
+        .with_retry_interval(retry_connection_interval)
+}
